TownySpawnUI 1.2.0 Changelog
+ Complete rewrite of how the nations are grabbed and put in the menu for it to be more efficient and less RAM eating
+ Updated for Towny 0.98.4.0 (Older versions will still work as intended)
+ Block names in /town set menu-item and /nation set menu-item can now be lowercase!

The pages are now made automatically and there's no more limits!
(Yes, there used to be a limit for how much pages of nations you could have, it was 3)

This is by far the update that took me the longest, so please report any bugs you find in the Discussions tab.

Hope you enjoy!